Added a "--pri-heur-topN-per" option to specify topN reads as a percentage instead of the raw number.  As always, existence of this parameter can be determined using the "--help" option in peace as in:
$ peace --heuristics primes --help

+++ peace/src/clustering/MSTClusterMaker.h
@@ -310,7 +310,7 @@ protected:
         be changed via the command-line argument \c --maxUse \e
         threshold.
     */
-    int maxUse;
+    float maxUse;
     
     /** Command line option to set the clustering threshold to be
         used in deriving clusters from the MST.
